Note api routes

+++ src/repositories/AppRepository.ts
@@ -8,6 +8,7 @@ import {
   GetEventsForUserDBResponse,
   ProcessedEvent,
   GetOrderDBResult,
+  DBNoteResult,
 } from "./AppRepository.types";
 import { AddEventBody, GenericSearchQuery, UpdateEventBody } from "../controllers/AppController.types";
 import { ContainerCradle } from "../container.types";
@@ -41,7 +42,7 @@ class AppRepository {
   }
 
   async getCountries({ searchTerm, offset, limit }: GenericSearchQuery) {
-    let query = this.knex<Country>("chrome_event_countdown").select("id", "name").from("countries").orderBy("name");
+    let query = this.knex("countries").select("id", "name").orderBy("name");
     query = this._applyOptionstoQuery(query, { searchTerm, offset, limit });
 
     const results = await this.db.runQuery<Country[]>({
@@ -52,9 +53,8 @@ class AppRepository {
   }
 
   async getCitiesForCountry({ countryId, searchTerm, offset, limit }: GetCitiesForCountryParams) {
-    let query = this.knex<City>("chrome_event_countdown")
+    let query = this.knex<City>("cities")
       .select("id", "name", "timezoneName")
-      .from("cities")
       .where("countryId", countryId)
       .orderBy("name");
 
@@ -68,10 +68,7 @@ class AppRepository {
   }
 
   async getCityCoordinates(cityId: number) {
-    const expression = this.knex<Coordinates>("chrome_event_countdown")
-      .select("lat", "lng")
-      .from("cities")
-      .where("id", cityId);
+    const expression = this.knex("cities").select("lat", "lng").where("id", cityId);
 
     const [cordinates] = await this.db.runQuery<Coordinates[]>({
       query: expression.toQuery(),
@@ -104,7 +101,7 @@ class AppRepository {
   }
 
   async getEventsForUser(userId: number) {
-    const expression = this.knex("chrome_event_countdown")
+    const expression = this.knex("events AS e")
       .select(
         "e.id",
         "e.name",
@@ -117,7 +114,6 @@ class AppRepository {
         "cu.id as countryId",
         "cu.name as countryName"
       )
-      .from("events AS e")
       .leftJoin("cities AS ci", "ci.id", "e.cityId")
       .leftJoin("countries AS cu", "cu.id", "ci.countryId")
       .where("e.userId", userId);
@@ -130,7 +126,7 @@ class AppRepository {
   }
 
   async getEvent(userId: number, eventId: number) {
-    const expression = this.knex("chrome_event_countdown")
+    const expression = this.knex("events AS e")
       .select(
         "e.id",
         "e.name",
@@ -143,7 +139,6 @@ class AppRepository {
         "cu.id as countryId",
         "cu.name as countryName"
       )
-      .from("events AS e")
       .leftJoin("cities AS ci", "ci.id", "e.cityId")
       .leftJoin("countries AS cu", "cu.id", "ci.countryId")
       .where("e.userId", userId)
@@ -168,7 +163,7 @@ class AppRepository {
   async addEvent(userId: number, event: AddEventBody) {
     const order = await this.getNextEventOrderForUser(userId);
 
-    const expression = this.knex("chrome_event_countdown").table("events").insert({
+    const expression = this.knex("events").insert({
       name: event.name,
       dateTime: event.eventDate,
       background: event.background,
@@ -187,8 +182,7 @@ class AppRepository {
   }
 
   async updateEvent(userId: number, eventId: number, event: UpdateEventBody) {
-    const expression = this.knex("chrome_event_countdown")
-      .table("events")
+    const expression = this.knex("events")
       .update({
         name: event.name,
         dateTime: event.eventDate,
@@ -208,11 +202,7 @@ class AppRepository {
   }
 
   async deleteEvent({ userId, eventId }: { userId: number; eventId: number }) {
-    const expression = this.knex("chrome_event_countdown")
-      .table("events")
-      .delete()
-      .where("id", eventId)
-      .where("userId", userId);
+    const expression = this.knex("events").delete().where("id", eventId).where("userId", userId);
 
     const result = await this.db.runQuery<OkPacket>({
       query: expression.toQuery(),
@@ -225,7 +215,7 @@ class AppRepository {
 
   async reorderEvents({ userId, eventIds }: { userId: number; eventIds: number[] }) {
     const makeQuery = ({ order, eventId, userId }: { order: number; eventId: number; userId: number }) => {
-      return this.knex().table("events").update("order", order).where("id", eventId).where("userId", userId).toQuery();
+      return this.knex("events").update("order", order).where("id", eventId).where("userId", userId).toQuery();
     };
 
     // Not ideal, but it'll work for now
@@ -233,6 +223,52 @@ class AppRepository {
       await this.db.runQuery({
         query: makeQuery({ order: parseInt(idx, 10) + 1, eventId: eventIds[idx], userId }),
       });
+    }
+  }
+
+  async getNotes({ userId }: { userId: number }) {
+    const notes = await this.db.runQuery<DBNoteResult[]>({
+      query: this.knex("notes").select("id", "text").where("userId", userId).toQuery(),
+    });
+
+    return notes;
+  }
+
+  async getNote({ noteId, userId }: { noteId: number; userId: number }) {
+    const [note] = await this.db.runQuery<DBNoteResult[]>({
+      query: this.knex("notes").select("id", "text").where("userId", userId).where("id", noteId).toQuery(),
+    });
+
+    return note;
+  }
+
+  async addNote({ text, userId }: { text: string; userId: number }) {
+    const result = await this.db.runQuery<OkPacket>({
+      query: this.knex("notes").insert({ text, userId }).toQuery(),
+    });
+
+    if (result.insertId) return result.insertId;
+
+    throw new Error("Failed to insert new note");
+  }
+
+  async updateNote({ noteId, text, userId }: { noteId: number; text: string; userId: number }) {
+    const result = await this.db.runQuery<OkPacket>({
+      query: this.knex("notes").update({ text }).where("id", noteId).where("userId", userId).toQuery(),
+    });
+
+    if (result.affectedRows !== 1) {
+      throw new Error("Failed to update note");
+    }
+  }
+
+  async deleteNote({ noteId, userId }: { noteId: number; userId: number }) {
+    const result = await this.db.runQuery<OkPacket>({
+      query: this.knex("notes").delete().where("id", noteId).where("userId", userId).toQuery(),
+    });
+
+    if (result.affectedRows !== 1) {
+      throw new Error("Failed to delete note");
     }
   }
 }
